Flatten Options
