+++ DevSkim-DotNet/Microsoft.DevSkim.VisualStudio/Options/OptionsPackage.cs
@@ -23,18 +23,8 @@
     [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
     [InstalledProductRegistration("#110", "#112", "1.0", IconResourceID = 400)] // Info on this package for Help/About
     [ProvideMenuResource("Menus.ctmenu", 1)]
-    [ProvideOptionPage(typeof(RuleOptionsPage), "DevSkim", "Rules", 1000, 1001, true)]
-    [ProvideProfile(typeof(RuleOptionsPage), "DevSkim", "Rules", 1000, 1002, true, DescriptionResourceID = 1003)]
-    [ProvideOptionPage(typeof(SuppressionOptionsPage), "DevSkim", "Suppressions", 1000, 1004, true)]
-    [ProvideProfile(typeof(SuppressionOptionsPage), "DevSkim", "Suppressions", 1000, 1005, true, DescriptionResourceID = 1006)]
-    [ProvideOptionPage(typeof(GuidanceOptionsPage), "DevSkim", "Guidance ", 1000, 1007, true)]
-    [ProvideProfile(typeof(GuidanceOptionsPage), "DevSkim", "Guidance", 1000, 1008, true, DescriptionResourceID = 1009)]
-    [ProvideOptionPage(typeof(IgnoreOptionsPage), "DevSkim", "Ignores", 1000, 1010, true)]
-    [ProvideProfile(typeof(IgnoreOptionsPage), "DevSkim", "Ignores", 1000, 1011, true, DescriptionResourceID = 1012)]
-    [ProvideOptionPage(typeof(FindingOptionsPage), "DevSkim", "Findings", 1000, 1013, true)]
-    [ProvideProfile(typeof(FindingOptionsPage), "DevSkim", "Findings", 1000, 1014, true, DescriptionResourceID = 1015)]
-    [ProvideOptionPage(typeof(TriggerOptionsPage), "DevSkim", "Triggers", 1000, 1016, true)]
-    [ProvideProfile(typeof(TriggerOptionsPage), "DevSkim", "Triggers", 1000, 1017, true, DescriptionResourceID = 1018)]
+    [ProvideOptionPage(typeof(GeneralOptionsPage), "DevSkim", "General", 1000, 1001, true)]
+    [ProvideProfile(typeof(GeneralOptionsPage), "DevSkim", "General", 1000, 1002, true, DescriptionResourceID = 1003)]
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1650:ElementDocumentationMustBeSpelledCorrectly", Justification = "pkgdef, VS and vsixmanifest are valid VS terms")]
     public sealed class OptionsPackage : AsyncPackage
     {
